docker/django/requirements.txt
django
psycopg[binary]
django-crispy-forms
crispy-bootstrap5
djangorestframework
